+++ src/SrUserEnrolment/EnrolmentWorkflow/Rule/MetaDataField/MetaDataField.php
@@ -44,7 +44,7 @@ class MetaDataField extends AbstractRule
                 return false;
 
             default:
-                return parent::supportsParentContext($parent_context);
+                return true;
         }
     }